% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_gmaData.R
\docType{data}
\name{data_mh_snp}
\alias{data_mh_snp}
\title{An example dataset for input}
\format{
a dataframe
}
\usage{
data_mh_snp
}
\description{
A dataset containing microhaps and SNPs for a large 
number of individuals across several populations
}
\keyword{datasets}
