% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falseGrandma.R
\name{falseGrandma}
\alias{falseGrandma}
\title{estimate per-comparison error rates}
\usage{
falseGrandma(gmaData, relationship = c("ssGP", "test1", "test2"),
  llrToTest, N = 10000, seed = NULL)
}
\arguments{
\item{gmaData}{the gmaData object containing your baseline populations and potential descendents. This 
input is created by \code{createGmaInput}}

\item{relationship}{the relationship you want to test for: "ssGP" - single sided grandparentage 
(a pair of either two maternal grandparents OR two paternal grandparents)}

\item{llrToTest}{a vector of llr thresholds to estimate false positive and false negative rates for}

\item{N}{the number of Monte Carlo samples to take to estimate the error rates}

\item{seed}{a positive integer to use as a seed for random number generation. If \code{NULL}, a seed is 
chosen based on the current system time}
}
\description{
some more here
}
\details{
and more here
}
