% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGrandma.R
\name{inferGrandma}
\alias{inferGrandma}
\title{Calculate log-likelihood ratios to infer relationships}
\usage{
inferGrandma(
  gmaData,
  relationship = c("ssGP", "sP"),
  crossRecords = NULL,
  minLLR = 0,
  filterLLR = TRUE,
  MIexcludeProb = 1e-04
)
}
\arguments{
\item{gmaData}{the gmaData object containing your baseline populations and potential descendents. This 
input is created by \code{createGmaInput}}

\item{relationship}{the relationship you want to test for: "ssGP" - single sided grandparentage 
(a pair of either two maternal grandparents OR two paternal grandparents)}

\item{crossRecords}{dataframe or matrix: column 1 is population, column 2 is individual
1 identifier, column 3 is individual 2 identifier}

\item{minLLR}{the minimum LLR to include in the results. Only used if \code{filterLLR} is \code{TRUE}.}

\item{filterLLR}{\code{TRUE} to filter results based on minLLR, \code{FALSE} to not.}

\item{MIexcludeProb}{the maximum probability of exclusion for a true relationship due to 
Mendelian incompatibilities. If \code{0}, then no filtering is performed
based on Mendelian incompatibilities.}
}
\description{
some more here
}
\details{
and some more here
}
