% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_gmaData.R
\name{print.gmaData}
\alias{print.gmaData}
\title{print method for gmaData}
\usage{
\method{print}{gmaData}(x, ...)
}
\arguments{
\item{x}{a gmaData object}

\item{...}{ignored}
}
\description{
print method for gmaData
}
