% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_gmaData.R
\name{createGmaInput}
\alias{createGmaInput}
\title{This creates a gmaData structure from user supplied dataframes}
\usage{
createGmaInput(baseline, mixture, unsampledPops = NULL,
  perSNPerror = NULL, dropoutProb = NULL)
}
\arguments{
\item{baseline}{a dataframe of the baseline individuals to use to infer relationships. The first column should
be the population each individual is from. The second column is the individual's identifier.
Following columns are genotypes, with columns 3 adn 4 being Allele1 and 2 for locus 1, columns
5 and 6 beign locus 2, ... This is a "two column per call" type of organization. Genotypes 
should be given as concatenated basecalls, with each base represented by a single character. Some examples of
microhaplotype alleles are "ACA", "AAD", "CTCTGGA". Some SNP alleles (which are just microhaplotypes
with a length of one base) are "A", "G", "D". In these examples, deletions are represented by "D". Missing genotypes
should be NA.}

\item{mixture}{a dataframe of the mixture individuals to infer relationships for. The first column  is the individual's identifier.
Following columns are genotypes, in the same manner as for \code{baseline}. The order and column names of the loci 
must be the same in both the baseline and mixture dataframes.}

\item{unsampledPops}{a dataframe of the individuals sampled from the "unsampled populations" used to estimate allele frequencies 
in these populations. See details for more explanation. Column 1 has the baseline population that an individual corresponds to, 
column 2 is the individual's identifier
(not currently used, but must be present). The following columns are genotypes, in the same manner as for \code{baseline}. 
The order and column names of the loci must be the same as in the baseline and mixture dataframes.}

\item{perSNPerror}{dataframe with each row representing a SNP. Column 1 is the locus name, column 2 is the order 
of the SNP in the locus (ie 1, 2, 3, ...), column 3 is the number of alleles at that SNP (not the locus as a whole),
and column 4 is the error rate (probability of observing any allele other than the correct allele at that SNP - not the 
locus as a whole).}

\item{dropoutProb}{dataframe with column 1 being locus name, column 2 being allele, and column 3 being dropout probability.
the locus name must be the column name of allele 1 for the corresponding locus in the baseline and mixture dataframes}
}
\description{
This creates a gmaData structure from user supplied dataframes
}
