% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falseGrandma.R
\name{falseGrandma}
\alias{falseGrandma}
\title{estimate per-comparison error rates}
\usage{
falseGrandma(
  gmaData,
  relationship = c("ssGP", "sP"),
  llrToTest,
  N = 10000,
  seed = NULL,
  itersPerMI = NULL,
  errorType = c("falseNegative", "pairwise", "Unrel", "Aunt", "HalfAunt", "ParCous")
)
}
\arguments{
\item{gmaData}{the gmaData object containing your baseline populations and potential descendents. This 
input is created by \code{createGmaInput}}

\item{relationship}{the relationship you want to test for: "ssGP" - single sided grandparentage 
(a pair of either two maternal grandparents OR two paternal grandparents)}

\item{llrToTest}{a vector of llr thresholds to estimate false positive and false negative rates for}

\item{N}{the number of Monte Carlo samples to take to estimate the error rates}

\item{seed}{a positive integer to use as a seed for random number generation. If \code{NULL}, a seed is 
chosen based on the current system time}

\item{pairwise}{TRUE to estimate false positive rate for pairwise combinations of the baseline populations.
In otherwords, the rate at which mixture individuals from one population have llr's equal to or above 
the threshold when being tested against the baseline of another population.}
}
\description{
some more here
}
\details{
and more here
}
