% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falseGrandma.R
\name{falseGrandma}
\alias{falseGrandma}
\title{estimate per-comparison error rates}
\usage{
falseGrandma(
  gmaData,
  relationship = c("ssGP", "sP"),
  llrToTest,
  N = 10000,
  seed = NULL,
  itersPerMI = NULL,
  errorType = c("falseNegative", "pairwise", "Unrel", "Aunt", "HalfAunt", "ParCous",
    "True_GAunt", "True_Unrel", "True_HGAunt", "True_GpCous", "GAunt_Unrel",
    "HGAunt_Unrel", "GpCous_Unrel", "GAunt", "GAunt_HGAunt", "Gaunt_GpCous", "HGAunt",
    "HGAunt_GpCous", "GpCous"),
  MIexcludeProb = 1e-04,
  maxMissingGenos = NULL,
  method = c("strat", "IS", "old")
)
}
\arguments{
\item{gmaData}{the gmaData object containing your baseline populations and potential descendents. This 
input is created by \code{createGmaInput}}

\item{relationship}{the relationship you want to test for: "ssGP" - single sided grandparentage 
(a pair of either two maternal grandparents OR two paternal grandparents); "sP" - single parent inference}

\item{llrToTest}{a vector of llr thresholds to estimate error rates for}

\item{N}{the number of Monte Carlo samples to take to estimate the error rates (ignored for stratified methods)}

\item{seed}{a positive integer to use as a seed for random number generation. If \code{NULL}, a seed is 
chosen based on the current system time}

\item{errorType}{the type of error estimate to make}

\item{MIexcludeProb}{the maximum probability of exclusion for a true relationship due to 
Mendelian incompatibilities. If \code{0}, then no filtering is performed
based on Mendelian incompatibilities.}

\item{maxMissingGenos}{the maximum number of missing genotypes a sample can have before you would 
choose to omit it from analysis}

\item{method}{strat for stratified, IS for importance sampling. Only used for ssGP. Do not use method = "old", 
this is currently for internal testing and will be soon removed.}

\item{itersperMI}{the number of iterations per Mendelian incompatibility, in order of 0, 1, ... 
(ignored for non-stratified methods)}
}
\description{
some more here
}
\details{
and more here
}
